Refactor abstract command into CRUD
