+++ src/main/java/seedu/address/logic/commands/contact/ClearContactCommand.java
@@ -2,7 +2,7 @@ package seedu.address.logic.commands.contact;
 
 import static seedu.address.logic.parser.CliSyntax.CONTACT_COMMAND_WORD;
 
-import seedu.address.logic.abstractcommand.ClearCommand;
+import seedu.address.logic.commands.delete.ClearCommand;
 import seedu.address.model.Model;
 import seedu.address.model.contact.Contact;
 import seedu.address.model.contact.ContactManager;
