+++ src/main/java/seedu/address/logic/commands/contact/DeleteContactCommand.java
@@ -5,7 +5,7 @@ import static seedu.address.logic.parser.CliSyntax.CONTACT_COMMAND_WORD;
 import seedu.address.commons.core.index.Index;
 import seedu.address.commons.util.ToStringBuilder;
 import seedu.address.logic.Messages;
-import seedu.address.logic.abstractcommand.DeleteCommand;
+import seedu.address.logic.commands.delete.DeleteCommand;
 import seedu.address.model.Model;
 import seedu.address.model.contact.Contact;
