+++ src/main/java/seedu/address/logic/commands/contact/ListContactCommand.java
@@ -2,7 +2,7 @@ package seedu.address.logic.commands.contact;
 
 import static seedu.address.logic.parser.CliSyntax.CONTACT_COMMAND_WORD;
 
-import seedu.address.logic.abstractcommand.ListCommand;
+import seedu.address.logic.commands.read.ListCommand;
 import seedu.address.model.Model;
 import seedu.address.model.contact.Contact;
