+++ src/main/java/seedu/address/logic/commands/todo/FilterTodoCommand.java
@@ -17,7 +17,7 @@ import seedu.address.commons.util.CollectionUtil;
 import seedu.address.commons.util.ToStringBuilder;
 import seedu.address.logic.Messages;
 import seedu.address.logic.TodoMessages;
-import seedu.address.logic.abstractcommand.FilterCommand;
+import seedu.address.logic.commands.read.FilterCommand;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Model;
 import seedu.address.model.contact.Contact;
